+++ CHANGES.rst
@@ -3,3 +3,5 @@ Changelog
 
 0.0-dev (unreleased)
 --------------------
+
+- Provide bower support through calmjs framework.